Make the conformance mirror 1:1 with Proofs/ + batch sweep reads
Mirror splits (the structure fence's model->home table is now one model
per module everywhere it isn't coupled by design):

- tooling_slots_queue_command.rs dissolved: the preflight/retry contract
  joins ToolExecution's home (tool_execution.rs); slot accounting ->
  fleet.rs; the command policy/sandbox/env contract -> command_policy.rs;
  the R4a queue/deadline rows project from Request/Session executables,
  so they join request_lifecycle.rs.
- transcript_background.rs split into transcript.rs and background.rs.
- New mcp_health.rs shape consumer pins the emitted threshold-projection
  case set from the conformance binary (incl. the H7 K=1
  healthy->evicted collapse and the K>=2 future rows); the transition
  behavior stays witnessed by the registered health_checker unit suite.
- New process.rs consumer pins the Process state machine completeness
  and the 25-case transition partition; the DB-backed drive stays
  registered in runtime_status::tests.
- PairingReconcile + ReversePairingHandlers documented as deliberately
  sharing one home (one two-node scenario harness exercises both).

All wrapper test names (= ledger consumer ids) unchanged, so no
ledger/registry churn.

Also: the periodic liveness sweep now loads every bridge child's
liveness row in ONE batched _in query per tick instead of one query per
bridge (review follow-up).

Filed #477 (generalize periodic sweeps behind a registry once a second
periodic sweep exists); closed #445 (verified sealed by
sanitize_history_for_provider + PromptAssembly fence + registered
boundary).
